Use most recent log, if any in analysis & Scoping
When we run an analysis (drop a file), if the file drop
contains no test or production log information, we will
reuse the previous valid test or production log information for
that specific application.  If it happens to be only Routes
that are dropped in, we don't bother analyzing right now

I also added some model level scoping.